ai_desc/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
deepl==1.12.0
Django==4.1.5
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
kombu==5.2.4
multidict==6.0.4
openai==0.26.0
prompt-toolkit==3.0.36
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.3.0
yarl==1.8.2
